fix: id change int to long

+++ app/src/main/java/com/exfarnanda1945/loremipsumgen/feat_history/domain/models/HistoryGenerator.kt
@@ -1,10 +1,7 @@
 package com.exfarnanda1945.loremipsumgen.feat_history.domain.models
 
-
-
-
 data class HistoryGenerator(
-    val id: Int,
+    val id: Long,
     val createdAt: Long,
     val result: String,
     val setting: GeneratorSetting
